compiler: fix support for non-utf8 terminals
The compiler echoes the names of the files being processed.
Those file names may contain UTF-8 characters.
However, some consoles may not support those.
This is notably the case of the Windows CMD or on Linux with the legacy
C locale. This caused the following error:

  ldgallery: <stdout>: commitBuffer: invalid argument (invalid character)

This patch adds automatic transliteration when a non-UTF8 terminal is
detected, so that neither the compiler or the terminal crash due to UTF8
characters.

GitHub: closes #341