viewer/command: add item download button
This adds a download button which allows the user to save the current
item as a file.

This is necessary because some item viewers do not expose a download
option on their own.

The download icon appears together with the other command buttons at the
top-left corner of the screen, replacing the listing sorting menu which
is only relevant for item lists (directory and search views).

GitHub: closes #308